feat: Implement Logger and Stats modules
Implemented Logger and Stats according to UML requirements.
Added thread-safe logging and implemented Stats constructors and fields.
Updated statistics display in Main.cpp.

+++ src/Bank.cpp
@@ -119,7 +119,7 @@ class Bank {
   }
 
   Stats getStats() {
-    int totalBalance = 0;
+    double totalBalance = 0.0;
 
     shared_lock lock(accountsMtx);
